feat(ui5-tabcontainer): Responsive paddings are now supported
Fixes: #2539
